feat: refactor+new_logic recommender

+++ backend/recommender/models.py
@@ -12,7 +12,7 @@ class SummaryViewLog(models.Model):
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     user_id = models.UUIDField(null=True, blank=True)
     summary_id = models.UUIDField()
-    duration_seconds = models.PositiveIntegerField()
+    duration_seconds = models.DecimalField(max_digits=10, decimal_places=2)
     viewed_at = models.DateTimeField(auto_now_add=True)
     created_at = models.DateTimeField(default=timezone.now)
     updated_at = models.DateTimeField(auto_now=True)
